examples: use uintptr_t for USERPTR and drop dead prev read
camcap: cast the memalign'd buffer to uintptr_t (add stdint.h) instead of unsigned long when setting buf.m.userptr.

jpegenc: remove the unused RAM-marker read (uint32_t prev = *PREV_MARK_ADDR) that only existed to be silenced with (void)prev.

+++ examples/camcap/camcap_main.c
@@ -7,6 +7,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 #include <fcntl.h>
 #include <unistd.h>
 #include <errno.h>
@@ -87,7 +88,7 @@ int main(int argc, char *argv[])
       buf.type      = type;
       buf.memory    = V4L2_MEMORY_USERPTR;
       buf.index     = i;
-      buf.m.userptr = (unsigned long)bufs[i];
+      buf.m.userptr = (uintptr_t)bufs[i];
       buf.length    = framesize;
       if (xioctl(fd, VIDIOC_QBUF, &buf) < 0)
         { printf("ERROR: QBUF %d: %d\n", i, errno); goto out; }
